Lift Update
- Added a Shack building for players to spawn in above the Cave. Shack contains the Elevator Lift and a new animated Steam Engine object that helps indicate the vertical position of the Lift.
- Added a LocalPlayerTracker script that attaches a trigger to the player in order to track players' collision with the Lift.

TODO:
Due to the speed the Lift has to travel at, players frequently have to catch up with the lift when going down, and will phase through the lift partway through when going up. To combat this, Stations should be used but it appears that setting them to Immobilize is the only way, which may require a rebuild of the Character Controller. More research is needed.