Audio Note
I realized last time I was messing with audio that the 80% slider didn't actually apply unless you pressed its play button, so I placeboed all the slow effects I was applying last time...

I tested slowness, and found that 80% was a pretty good level to drop to which in turn adds a whopping 20% to the length of each clip. The clips will need to be re-exported, but first the exact fade in/fade out settings must be identified (at a later date).